su tests: removed VC warnings from test_su.c, torture_su_alloc.c, torture_su_bm.c, torture_su_time.c.
Patch from Michael Jerris.

darcs-hash:20061122173712-65a35-9125e1ba2192103096c17b73fe4dd23b27da5cf7.gz

+++ libsofia-sip-ua/su/torture_su_bm.c
@@ -176,7 +176,7 @@ char const Needle[] =
 "ULLAMCORPER PRETIUM LOREM. VESTIBULUM ANTE IPSUM PRIMIS IN FAUCIBUS ORCI "
 "LUCTUS ET ULTRICES POSUERE CUBILIA CURAE;\n";
 
-    int nlen = strlen(needle);
+    size_t nlen = strlen(needle);
 
     bm_fwd_table_t *fwd;
